+++ src/app/(sign)/@agreement/register/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import clsx from 'clsx';
 import { useModalStore } from '@/store/userStore';
-import AgreementForm from '../../_component/AgreementForm';
+import AgreementForm from '../_component/AgreementForm';
 
 export default function LoginModal() {
   const modalState = useModalStore();
