+++ src/app/(sign)/_component/RegisterForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useValidate } from '@/store/userStore';
-import ProfilForm from './ProfilForm';
-import UserInfoForm from './UserInfoForm';
+import ProfilForm from './_RegisterForm/ProfilForm';
+import UserInfoForm from './_RegisterForm/UserInfoForm';
 
 export default function RegisterForm() {
   const validate = useValidate();
